- Fixed some warings usng LLVM 3.3 on XCode 4.2.1.

+++ include/cybergarage/util/ctime.h
@@ -36,7 +36,7 @@ extern "C" {
 ****************************************/
 
 #if defined(BTRON) || defined(ITRON) 
-typedef long CgSysTime;
+typedef size_t CgSysTime;
 #else
 typedef time_t CgSysTime;
 #endif
